requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15.1     
charset-normalizer==2.1.1
idna==3.3
Pillow==9.2.0
pony==0.7.16
psycopg2==2.9.3
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
vk-api==11.9.9
